+++ Pal/src/host/Linux-SGX/db_streams.c
@@ -319,7 +319,7 @@ int _DkReceiveHandle(PAL_HANDLE hdl, PAL_HANDLE * cargo)
 
     if (IS_ERR(ret))
         return unix_to_pal_error(ERRNO(ret));
-    if (ret < sizeof(struct hdl_header)) {
+    if ((size_t)ret < sizeof(struct hdl_header)) {
         /*
          * This code block is just in case to cover all the possibilities
          * to shield Iago attack.
@@ -343,13 +343,13 @@ int _DkReceiveHandle(PAL_HANDLE hdl, PAL_HANDLE * cargo)
 
     // initialize variables to get body
     void * buffer = __alloca(hdl_hdr.data_size);
-    unsigned int nfds = 0;
+    uint32_t nfds = 0;
 
     for (int i = 0 ; i < MAX_FDS ; i++)
         if (hdl_hdr.fds & (1U << i))
             nfds++;
 
-    unsigned int * fds = __alloca(sizeof(unsigned int) * nfds);
+    uint32_t * fds = __alloca(sizeof(unsigned int) * nfds);
 
     ret = ocall_sock_recv_fd(ch, buffer, hdl_hdr.data_size,
                              fds, &nfds);
@@ -362,8 +362,8 @@ int _DkReceiveHandle(PAL_HANDLE hdl, PAL_HANDLE * cargo)
     if (ret < 0)
         return ret;
 
-    int n = 0;
-    for (int i = 0 ; i < MAX_FDS ; i++)
+    uint32_t n = 0;
+    for (uint32_t i = 0 ; i < MAX_FDS ; i++)
         if (hdl_hdr.fds & (1U << i)) {
             if (n < nfds) {
                 handle->generic.fds[i] = fds[n++];
